+++ adyenv6core/testsrc/com/adyen/v6/service/AdyenNotificationServiceTest.java
@@ -1,11 +1,19 @@
-package com.adyen.v6.cronjob;
+package com.adyen.v6.service;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.runners.MockitoJUnitRunner;
 import com.adyen.v6.model.NotificationItemModel;
-import com.adyen.v6.repository.NotificationItemRepository;
 import com.adyen.v6.repository.OrderRepository;
 import com.adyen.v6.repository.PaymentTransactionRepository;
-import com.adyen.v6.service.AdyenBusinessProcessService;
-import com.adyen.v6.service.AdyenTransactionService;
 import de.hybris.bootstrap.annotations.UnitTest;
 import de.hybris.platform.core.model.order.OrderModel;
 import de.hybris.platform.orderprocessing.model.OrderProcessModel;
@@ -15,25 +23,15 @@ import de.hybris.platform.processengine.BusinessProcessService;
 import de.hybris.platform.returns.model.ReturnProcessModel;
 import de.hybris.platform.returns.model.ReturnRequestModel;
 import de.hybris.platform.servicelayer.model.ModelService;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.runners.MockitoJUnitRunner;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import static com.adyen.model.notification.NotificationRequestItem.*;
+import static com.adyen.model.notification.NotificationRequestItem.EVENT_CODE_AUTHORISATION;
+import static com.adyen.model.notification.NotificationRequestItem.EVENT_CODE_CAPTURE;
+import static com.adyen.model.notification.NotificationRequestItem.EVENT_CODE_REFUND;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @UnitTest
 @RunWith(MockitoJUnitRunner.class)
-public class AdyenProcessNotificationCronJobTest {
+public class AdyenNotificationServiceTest {
     @Mock
     private ModelService modelServiceMock;
 
@@ -47,33 +45,22 @@ public class AdyenProcessNotificationCronJobTest {
     private PaymentTransactionRepository paymentTransactionRepositoryMock;
 
     @Mock
-    private NotificationItemRepository notificationItemRepositoryMock;
-
-    private AdyenBusinessProcessService adyenBusinessProcessService;
-
-    @Mock
     private BusinessProcessService businessProcessServiceMock;
 
     @Mock
     private PaymentTransactionEntryModel paymentTransactionEntryModelMock;
 
-    private AdyenProcessNotificationCronJob adyenProcessNotificationCronJob;
+    @InjectMocks
+    private DefaultAdyenNotificationService adyenNotificationService;
 
     @Before
     public void setUp() {
-        when(modelServiceMock.create(PaymentTransactionEntryModel.class))
-                .thenReturn(new PaymentTransactionEntryModel());
+        when(modelServiceMock.create(PaymentTransactionEntryModel.class)).thenReturn(new PaymentTransactionEntryModel());
 
-        adyenBusinessProcessService = new AdyenBusinessProcessService();
+        DefaultAdyenBusinessProcessService adyenBusinessProcessService = new DefaultAdyenBusinessProcessService();
         adyenBusinessProcessService.setBusinessProcessService(businessProcessServiceMock);
 
-        adyenProcessNotificationCronJob = new AdyenProcessNotificationCronJob();
-        adyenProcessNotificationCronJob.setModelService(modelServiceMock);
-        adyenProcessNotificationCronJob.setAdyenBusinessProcessService(adyenBusinessProcessService);
-        adyenProcessNotificationCronJob.setAdyenTransactionService(adyenTransactionServiceMock);
-        adyenProcessNotificationCronJob.setOrderRepository(orderRepositoryMock);
-        adyenProcessNotificationCronJob.setNotificationItemRepository(notificationItemRepositoryMock);
-        adyenProcessNotificationCronJob.setPaymentTransactionRepository(paymentTransactionRepositoryMock);
+        adyenNotificationService.setAdyenBusinessProcessService(adyenBusinessProcessService);
     }
 
     @After
@@ -84,8 +71,6 @@ public class AdyenProcessNotificationCronJobTest {
     /**
      * Test CAPTURE notification handling
      * It should save the capture notification and emmit the AdyenCaptured event
-     *
-     * @throws Exception
      */
     @Test
     public void testCaptureNotification() throws Exception {
@@ -96,8 +81,7 @@ public class AdyenProcessNotificationCronJobTest {
 
         paymentTransactionModel.setOrder(orderModel);
 
-        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class)))
-                .thenReturn(paymentTransactionModel);
+        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class))).thenReturn(paymentTransactionModel);
 
         NotificationItemModel notificationItemModel = new NotificationItemModel();
         notificationItemModel.setOriginalReference("123");
@@ -105,12 +89,9 @@ public class AdyenProcessNotificationCronJobTest {
         notificationItemModel.setEventCode(EVENT_CODE_CAPTURE);
         notificationItemModel.setSuccess(true);
 
-        when(adyenTransactionServiceMock.createCapturedTransactionFromNotification(
-                paymentTransactionModel,
-                notificationItemModel
-        )).thenReturn(paymentTransactionEntryModelMock);
+        when(adyenTransactionServiceMock.createCapturedTransactionFromNotification(paymentTransactionModel, notificationItemModel)).thenReturn(paymentTransactionEntryModelMock);
 
-        adyenProcessNotificationCronJob.processNotification(notificationItemModel);
+        adyenNotificationService.processNotification(notificationItemModel);
 
         //Verify that we emmit the event of Capture to the order processes
         verify(businessProcessServiceMock).triggerEvent("order_process_code_AdyenCaptured");
@@ -122,8 +103,6 @@ public class AdyenProcessNotificationCronJobTest {
     /**
      * Test AUTHORISATION notification handling
      * It should save the capture notification and emmit the AdyenCaptured event
-     *
-     * @throws Exception
      */
     @Test
     public void testAuthorisationNotification() throws Exception {
@@ -138,13 +117,11 @@ public class AdyenProcessNotificationCronJobTest {
         notificationItemModel.setMerchantReference(merchantReference);
         notificationItemModel.setSuccess(true);
 
-        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class)))
-                .thenReturn(null);
+        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class))).thenReturn(null);
 
-        when(orderRepositoryMock.getOrderModel(Mockito.any(String.class)))
-                .thenReturn(orderModel);
+        when(orderRepositoryMock.getOrderModel(Mockito.any(String.class))).thenReturn(orderModel);
 
-        adyenProcessNotificationCronJob.processNotification(notificationItemModel);
+        adyenNotificationService.processNotification(notificationItemModel);
 
         //Verify that we emmit the event of Capture to the order processes
         verify(businessProcessServiceMock).triggerEvent("order_process_code_AdyenAuthorized");
@@ -169,13 +146,11 @@ public class AdyenProcessNotificationCronJobTest {
         notificationItemModel.setMerchantReference(merchantReference);
         notificationItemModel.setSuccess(false);
 
-        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class)))
-                .thenReturn(null);
+        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class))).thenReturn(null);
 
-        when(orderRepositoryMock.getOrderModel(Mockito.any(String.class)))
-                .thenReturn(orderModel);
+        when(orderRepositoryMock.getOrderModel(Mockito.any(String.class))).thenReturn(orderModel);
 
-        adyenProcessNotificationCronJob.processNotification(notificationItemModel);
+        adyenNotificationService.processNotification(notificationItemModel);
 
         //Verify that we emmit the event of Capture to the order processes
         verify(businessProcessServiceMock).triggerEvent("order_process_code_AdyenAuthorized");
@@ -194,10 +169,9 @@ public class AdyenProcessNotificationCronJobTest {
         notificationItemModel.setEventCode(EVENT_CODE_AUTHORISATION);
         notificationItemModel.setSuccess(true);
 
-        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class)))
-                .thenReturn(new PaymentTransactionModel());
+        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class))).thenReturn(new PaymentTransactionModel());
 
-        adyenProcessNotificationCronJob.processNotification(notificationItemModel);
+        adyenNotificationService.processNotification(notificationItemModel);
 
         //Verify that we emmit the event of Capture to the order processes
         verify(orderRepositoryMock, Mockito.never()).getOrderModel(Mockito.any(String.class));
@@ -205,8 +179,6 @@ public class AdyenProcessNotificationCronJobTest {
 
     /**
      * Test successful Refund notification
-     *
-     * @throws Exception
      */
     @Test
     public void testRefundNotification() throws Exception {
@@ -217,8 +189,7 @@ public class AdyenProcessNotificationCronJobTest {
 
         paymentTransactionModel.setOrder(orderModel);
 
-        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class)))
-                .thenReturn(paymentTransactionModel);
+        when(paymentTransactionRepositoryMock.getTransactionModel(Mockito.any(String.class))).thenReturn(paymentTransactionModel);
 
         NotificationItemModel notificationItemModel = new NotificationItemModel();
         notificationItemModel.setOriginalReference("123");
@@ -226,12 +197,9 @@ public class AdyenProcessNotificationCronJobTest {
         notificationItemModel.setEventCode(EVENT_CODE_REFUND);
         notificationItemModel.setSuccess(true);
 
-        when(adyenTransactionServiceMock.createRefundedTransactionFromNotification(
-                paymentTransactionModel,
-                notificationItemModel
-        )).thenReturn(paymentTransactionEntryModelMock);
+        when(adyenTransactionServiceMock.createRefundedTransactionFromNotification(paymentTransactionModel, notificationItemModel)).thenReturn(paymentTransactionEntryModelMock);
 
-        adyenProcessNotificationCronJob.processNotification(notificationItemModel);
+        adyenNotificationService.processNotification(notificationItemModel);
 
         //Verify that we emmit the event of Capture to the order processes
         verify(businessProcessServiceMock).triggerEvent("return_process_code_AdyenRefunded");
